+++ src/core_modules/capture-core/components/TeiSearch/epics/teiSearch.epics.js
@@ -36,7 +36,7 @@ import type { QuerySingleResource } from '../../../utils/api/api.types';
 const getOuQueryArgs = (orgUnit: ?Object, orgUnitScope: string) =>
     (['ACCESSIBLE', 'CAPTURE', 'ALL'].includes(orgUnitScope) ?
         { ouMode: orgUnitScope } :
-        { ou: orgUnit && orgUnit.id, ouMode: orgUnitScope });
+        { orgUnit: orgUnit && orgUnit.id, ouMode: orgUnitScope });
 
 const getContextQueryArgs = (programId: ?string, trackedEntityTypeId: string) =>
     (programId ? { program: programId } : { trackedEntityType: trackedEntityTypeId });